feat(Slider.php): add Slider block component to Filament module
This commit adds a new file `Slider.php` to the `Filament/Blocks` directory. The `Slider` class is a block component that can be used in Filament forms. It provides a form input for selecting a method and a dropdown for selecting a layout. The `Slider` block component is configured to have 2 columns when used in a form context and 1 column otherwise.

The `Slider` block component also includes a view variable `$views` that is populated by executing the `GetViewsSiblingsAndSelfAction` action with the view `ui::components.blocks.slider.v1`. This action retrieves the sibling views and the view itself for the given view.

The `Slider` block component is intended to be used in the UI module and is namespaced under `Modules\UI\Filament\Blocks`.

feat(Slider.php): add paperino.blade.php, pippo.blade.php, and pluto.blade.php views for Slider block component

This commit adds three new blade view files `paperino.blade.php`, `pippo.blade.php`, and `pluto.blade.php` to the `Resources/views/components/blocks/slider` directory. These views are intended to be used as layouts for the `Slider` block component.